Use a much simpler (more code-sharing) algorithm for boundary coloring.
C-reduction is unchanged at the moment but will be implemented
next and the old/messy boundary-coloring code deleted.

The old one without C-reduction still drew edges between boundary
node zero and the rest, or (for configurations that have more than
one node) between every boundary node and an imaginary node in the
middle; the new one doesn't (at present, and probably shouldn't
when asking for regular boundary coloring).  The new one is also
better at not making redundant results that are the same except
third/fourth color swapped, because findColorings() has a better
algorithm for that (used when skipSameUpToColorRenaming = True).