Replace generatePossibleBoundaryColoringsTryingCReduction's impl
The new implementation:

 * Makes lots more sense.

 * Doesn't have enough (any!) tests specifically for
generatePossibleBoundaryColoringsGivenSmallerConfiguration() yet.

 * Currently does exactly the same thing, resulting in exactly the
same can/can't reduce (except that it generates colorings in a
different order and possibly with fewer redundant ones: neither have
an effect except on which config is chosen to show as a failing one,
which is why test.dot/test.png changed a little bit.)

 * Possibly generates exact duplicate colorings by coloring the
center node different ways (and then that node is not part of the
coloring report for the original boundary nodes). This is not
necessarily a problem but would be better not to be true.